LOTS
renaming old files, renaming the new files. coding as if basic_absorption_parameters exists and halfway through figuring out basic_absorption_parameters variable names and what they do.
