--- .absorption_code2021.py.swp
@@ -1,1 +1,0 @@
-VSCODE/c987b488f0159d4bff57462055d4feadf5e079b7dca009fdbfdacc2e9eaf8fca
