+++ utility_functions.py
@@ -2,6 +2,7 @@ import csv
 import numpy as np 
 from data_types import ColumnIndexes, RangesData, PointData
 import pandas as pd
+import scipy.constants as sc
 
 ######################################### sphinx ######################################### 
 """
@@ -21,17 +22,6 @@ def read_file(FILE: str):
             redshift_value_list.append(np.float(each_row_in_file[1]))
             snr_value_list.append(np.float(each_row_in_file[2]))
     return(redshift_value_list, snr_value_list, spectra_list)
-
-def read_file_abs(FILE: str):
-    spectra_index, spectra_filename, norm_spectra_filename, redshift_value, calc_snr_value, sdss_snr, bf, cf = [], [], [], [], [], [], [], []
-
-    with open(FILE) as f:  
-        for line in f:
-            each_row_in_file = line.split(",")
-            norm_spectra_filename.append(each_row_in_file[2])
-            redshift_value.append(np.float(each_row_in_file[3]))
-            calc_snr_value.append(np.float(each_row_in_file[4]))
-    return(redshift_value, calc_snr_value, norm_spectra_filename)
 
 def print_to_file(text: str, file_name: str):
     print(text, file = open(file_name, 'a'))
@@ -105,9 +95,66 @@ def read_list_spectra(file_name: str, column_list: list):
     return(spectra_list,redshift_list, snr_list)
 
 def read_spectra(spectra_data):
+    """Reads in and returns a lists of lists containing the wavelength, flux, and error for each spectra.
+
+    Defines the variables to be used in the code.
+
+    Parameters
+    ----------
+    spectra_data:
+        The drX (X being 9 or 16) files in the form of a text file conataining the wavelength, flux
+        and error of that paticular spectra.
+
+    Returns
+    -------
+    wavelength: list
+        All of the wavelength values in a list.
+    flux: list
+        All of the flux values in a list.
+    error: list
+        All of the error values in a list.
+
+    Note
+    ----
+    As shown in the return, the return value are lists within a list.
+    """
+
     column_index = ColumnIndexes(0, 1, 2)
     wavelength = spectra_data[:, column_index.wavelength]
     flux = spectra_data[:, column_index.flux] 
     error = spectra_data[:, column_index.error] 
     
     return [wavelength, flux, error]
+
+def wavelength_to_velocity(redshift, wavelength):
+    """Reads in a list of wavelength values to be converted to velocity.
+
+    Parameters
+    ----------
+    redshift: list
+        The list of redshift values needed for the equation.
+    wavelength: list
+        The list of wavelength values that will be converted to velocity.
+        
+    Returns
+    -------
+    beta: array
+        The values of velocity that were converted from wavelength.
+    """
+    # CIV doublet data from verner table
+    avr_CIV_doublet = 1549.0524
+
+    # Transform the wavelength array to velocity (called "beta") based on the CIV doublet: 
+    c_in_km = sc.speed_of_light**-3
+    z_absC = (wavelength / avr_CIV_doublet) - 1.
+    RC = (1. + redshift) / (1. + z_absC)
+    betaC = ((RC**2.) - 1.) / ((RC**2.) + 1.) # betaC is in units of c (speed of light)
+    betakm = -betaC * c_in_km #betakm is in km/s
+    beta = []
+
+    for velocity in betakm:
+        betas = round(velocity, 5)
+        beta.append(betas)
+    beta = np.array(beta)
+
+    return beta
